// src/types/player.ts
export interface Player {
  Player: string;
  Player_ID: string;
  Country: string;
  Player_Role: string;
  Role_Detail: string;
  Birth_Date: string;
  Birth_Place: string;
  Height: string;
  Batting_Style: string;
  Bowling_Style: string;
  Team_Name: string;
  Team_ID: string;
  Price: number;
  // Add other fields as needed
} 